# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "РезюмИТ Найм",
    "version": "1.4.0"
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "summary": "Jobs",
        "operationId": "jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Job",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/applications": {
      "get": {
        "summary": "My Applications",
        "operationId": "my_applications_api_applications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobBody": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 3,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 30,
            "title": "Description"
          },
          "terms": {
            "type": "string",
            "maxLength": 500,
            "title": "Terms",
            "default": ""
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "type": "array",
            "maxItems": 15,
            "minItems": 1,
            "title": "Requirements"
          },
          "screening_questions": {
            "items": {
              "type": "string",
              "enum": [
                "screen_motivation",
                "screen_conditions",
                "screen_availability"
              ]
            },
            "type": "array",
            "maxItems": 3,
            "title": "Screening Questions"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "requirements"
        ],
        "title": "JobBody"
      },
...
      "Requirement": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,40}$",
            "title": "Id"
          },
          "skill": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Skill"
          },
          "label": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Label"
          },
          "type": {
            "type": "string",
            "enum": [
              "must",
              "nice"
            ],
            "title": "Type",
            "default": "must"
          },
          "source": {
            "type": "string",
            "maxLength": 700,
            "title": "Source",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "skill",
          "label"
        ],
        "title": "Requirement"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "IntegrationKey": {
        "type": "http",
        "description": "Scoped rhi_ key, created by employer/admin. Not a JWT. Server-side only.",
        "scheme": "bearer"
      }
    }
  }
}